MAINT: Moved `request_analysis` to controller.py
